Updated all the OpenTK references to use NuGet (#4836)
* Updated all the OpenTK references to use NuGet
 - Using the same NuGet version of OpenTK across the board
 - Added a gallery factory to allow for platform-specific pages
 - Added the OpenGL galleries to Android, iOS, GTK# and WPF

* Fix the build for iOS and macOS

* Fix the modified defines for the control gallery
