+++ .nuspec/Xamarin.Forms.Platform.WPF.nuspec
@@ -15,8 +15,8 @@
     <dependencies>
       <group targetFramework="net45">
         <dependency id="Xamarin.Forms" version="$version$"/>
-        <dependency id="OpenTK" version="2.0.0"/>
-        <dependency id="OpenTK.GLControl" version="1.1.2349.61993"/>
+        <dependency id="OpenTK" version="3.0.1"/>
+        <dependency id="OpenTK.GLControl" version="3.0.1"/>
       </group>
     </dependencies>
   </metadata>
