--- Xamarin.Forms.ControlGallery.GTK/StringProvider.cs
@@ -1,17 +1,0 @@
-﻿using Xamarin.Forms;
-using Xamarin.Forms.ControlGallery.GTK;
-
-[assembly: Dependency(typeof(StringProvider))]
-namespace Xamarin.Forms.ControlGallery.GTK
-{
-    public class StringProvider : Controls.IStringProvider
-    {
-        public string CoreGalleryTitle
-        {
-            get
-            {
-                return "GTK# CoreGallery";
-            }
-        }
-    }
-}
