+++ Xamarin.Forms.ControlGallery.MacOS/PlatformSpecificCoreGalleryFactory.cs
@@ -1,0 +1,25 @@
+﻿using System;
+using System.Collections.Generic;
+using Xamarin.Forms;
+using Xamarin.Forms.ControlGallery.MacOS;
+using Xamarin.Forms.Controls;
+
+[assembly: Dependency(typeof(PlatformSpecificCoreGalleryFactory))]
+
+namespace Xamarin.Forms.ControlGallery.MacOS
+{
+	public class PlatformSpecificCoreGalleryFactory : IPlatformSpecificCoreGalleryFactory
+	{
+		public string Title => "macOS Core Gallery";
+
+		public IEnumerable<(Func<Page> Create, string Title)> GetPages()
+		{
+#if HAVE_OPENTK
+			yield return (() => new BasicOpenGLGallery(), "Basic OpenGL Gallery - Legacy");
+			yield return (() => new AdvancedOpenGLGallery(), "Advanced OpenGL Gallery - Legacy");
+#else
+			return null;
+#endif
+		}
+	}
+}
