+++ Xamarin.Forms.ControlGallery.Tizen/PlatformSpecificCoreGalleryFactory.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using Xamarin.Forms;
+using Xamarin.Forms.ControlGallery.Tizen;
+using Xamarin.Forms.Controls;
+
+[assembly: Dependency(typeof(PlatformSpecificCoreGalleryFactory))]
+
+namespace Xamarin.Forms.ControlGallery.Tizen
+{
+	public class PlatformSpecificCoreGalleryFactory : IPlatformSpecificCoreGalleryFactory
+	{
+		public string Title => "Tizen Core Gallery";
+
+		public IEnumerable<(Func<Page> Create, string Title)> GetPages() => null;
+	}
+}
