+++ Xamarin.Forms.ControlGallery.WindowsUniversal/PlatformSpecificCoreGalleryFactory.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using Xamarin.Forms;
+using Xamarin.Forms.ControlGallery.WindowsUniversal;
+using Xamarin.Forms.Controls;
+
+[assembly: Dependency(typeof(PlatformSpecificCoreGalleryFactory))]
+
+namespace Xamarin.Forms.ControlGallery.WindowsUniversal
+{
+	public class PlatformSpecificCoreGalleryFactory : IPlatformSpecificCoreGalleryFactory
+	{
+		public string Title => "Windows Core Gallery";
+
+		public IEnumerable<(Func<Page> Create, string Title)> GetPages() => null;
+	}
+}
