--- Xamarin.Forms.ControlGallery.WindowsUniversal/StringProvider.cs
@@ -1,15 +1,0 @@
-﻿using Xamarin.Forms;
-using Xamarin.Forms.ControlGallery.WindowsUniversal;
-using Xamarin.Forms.Controls;
-
-[assembly: Dependency (typeof (StringProvider))]
-
-namespace Xamarin.Forms.ControlGallery.WindowsUniversal
-{
-	public class StringProvider : IStringProvider
-	{
-		public string CoreGalleryTitle {
-			get { return "Windows Core Gallery"; }
-		}
-	}
-}
