+++ Xamarin.Forms.Controls/CoreGallery.cs
@@ -370,7 +370,8 @@ namespace Xamarin.Forms.Controls
 				new GalleryPageFactory(() => new MultiGallery(), "Multi Gallery - Legacy"),
 				new GalleryPageFactory(() => new NavigationPropertiesGallery(), "Navigation Properties"),
 #if HAVE_OPENTK
-				new GalleryPageFactory(() => new OpenGLGallery(), "OpenGLGallery - Legacy"),
+				new GalleryPageFactory(() => new BasicOpenGLGallery(), "Basic OpenGL Gallery - Legacy"),
+				new GalleryPageFactory(() => new AdvancedOpenGLGallery(), "Advanced OpenGL Gallery - Legacy"),
 #endif
 				new GalleryPageFactory(() => new PickerGallery(), "Picker Gallery - Legacy"),
 				new GalleryPageFactory(() => new ProgressBarGallery(), "ProgressBar Gallery - Legacy"),
@@ -396,6 +397,12 @@ namespace Xamarin.Forms.Controls
 
 		public CorePageView(Page rootPage, NavigationBehavior navigationBehavior = NavigationBehavior.PushAsync)
 		{
+			var galleryFactory = DependencyService.Get<IPlatformSpecificCoreGalleryFactory>();
+
+			var platformPages = galleryFactory?.GetPages();
+			if (platformPages != null)
+				_pages.AddRange(platformPages.Select(p => new GalleryPageFactory(p.Create, p.Title + " (Platform Specifc)")));
+
 			_titleToPage = _pages.ToDictionary(o => o.Title);
 
 			// avoid NRE for root pages without NavigationBar
@@ -492,9 +499,9 @@ namespace Xamarin.Forms.Controls
 		{
 			ValidateRegistrar();
 
-			IStringProvider stringProvider = DependencyService.Get<IStringProvider>();
+			var galleryFactory = DependencyService.Get<IPlatformSpecificCoreGalleryFactory>();
 
-			Title = stringProvider.CoreGalleryTitle;
+			Title = galleryFactory?.Title ?? "Core Gallery";
 
 			var corePageView = new CorePageView(rootPage, navigationBehavior);
 
@@ -568,9 +575,11 @@ namespace Xamarin.Forms.Controls
 	}
 
 	[Preserve(AllMembers = true)]
-	public interface IStringProvider
+	public interface IPlatformSpecificCoreGalleryFactory
 	{
-		string CoreGalleryTitle { get; }
+		string Title { get; }
+
+		IEnumerable<(Func<Page> Create, string Title)> GetPages();
 	}
 	[Preserve(AllMembers = true)]
 	public static class CoreGallery
