+++ Xamarin.Forms.Controls/RootPages/RootTabbedNavigationContentPage.cs
@@ -5,10 +5,6 @@ using System.Linq;
 using System.Resources;
 using System.Text;
 using System.Threading.Tasks;
-#if HAVE_OPENTK
-using OpenTK.Graphics;
-using OpenTK.Graphics.ES20;
-#endif
 
 namespace Xamarin.Forms.Controls
 {
