--- Xamarin.Forms.Platform.GTK/Libs/OpenTK/README.txt
@@ -1,1 +1,0 @@
-﻿https://github.com/opentk/opentk
